Attach id buffers to fragment shaders, and transfer buffers back to cpu.
This provides the bulk of what is needed for mouse picking. It's is
still a proof-of-concept though - to finalize mouse picking we need to
pass in the ids for each entity to the vertex shader, pass it through to
the fragment shader, and finally display _that_ id instead of the
hard-coded "1" we're using now.